✨ Refactor error handling: consolidate error classes and update exception handlers
