--- services/catalog/src/simcore_service_catalog/exceptions/handlers/__init__.py
@@ -1,24 +1,0 @@
-from fastapi import FastAPI, HTTPException, status
-from fastapi.exceptions import RequestValidationError
-
-from ._http_error import http_error_handler, make_http_error_handler_for_exception
-from ._validation_error import http422_error_handler
-
-
-def setup_exception_handlers(app: FastAPI) -> None:
-    app.add_exception_handler(HTTPException, http_error_handler)
-    app.add_exception_handler(RequestValidationError, http422_error_handler)
-
-    # SEE https://docs.python.org/3/library/exceptions.html#exception-hierarchy
-    app.add_exception_handler(
-        NotImplementedError,
-        make_http_error_handler_for_exception(
-            status.HTTP_501_NOT_IMPLEMENTED, NotImplementedError
-        ),
-    )
-    app.add_exception_handler(
-        Exception,
-        make_http_error_handler_for_exception(
-            status.HTTP_500_INTERNAL_SERVER_ERROR, Exception
-        ),
-    )
